requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.4
numpy==1.26.3
xgboost==2.0.3
lightgbm==4.3.0
catboost==1.2.2
scikit-learn==1.4.0
pydantic==2.5.3
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0
validators==0.22.0
tqdm==4.67.1
